requirements.txt
streamlit
httpx-oauth
python-dotenv